ejercicio carga de modelos OK, codigo de ejercicios anteriores comentado/borrado y mejora en el control de la camara (ahora se dirige bien a donde apunta el vector direccion)

+++ Practica1/src/Camera.h
@@ -16,8 +16,8 @@ public:
 	glm::mat4 LookAt();
 	GLfloat GetFOV();
 	GLboolean moveForward, moveBackwards, moveLeft, moveRight;
-	double Deltatime;
-	double Lastframe;
+	float Deltatime;
+	float Lastframe;
 	~Camera();
 private:
 	glm::vec3 cameraPos;
